+++ cvat/apps/engine/signals.py
@@ -1,14 +1,25 @@
 # Copyright (C) 2019 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
+import shutil
 
-from .models import Job, StatusChoice
+from django.db.models.signals import post_delete, post_save
+from django.dispatch import receiver
 
+from .models import (
+    Data,
+    Job,
+    StatusChoice,
+    Task,
+)
+
+
+@receiver(post_save, sender=Job, dispatch_uid="update_task_status")
 def update_task_status(instance, **kwargs):
     db_task = instance.segment.task
     db_jobs = list(Job.objects.filter(segment__task_id=db_task.id))
     status = StatusChoice.COMPLETED
-    if   list(filter(lambda x: x.status == StatusChoice.ANNOTATION, db_jobs)):
+    if list(filter(lambda x: x.status == StatusChoice.ANNOTATION, db_jobs)):
         status = StatusChoice.ANNOTATION
     elif list(filter(lambda x: x.status == StatusChoice.VALIDATION, db_jobs)):
         status = StatusChoice.VALIDATION
@@ -17,3 +28,12 @@ def update_task_status(instance, **kwargs):
         db_task.status = status
         db_task.save()
 
+
+@receiver(post_delete, sender=Task, dispatch_uid="delete_task_files_on_delete_task")
+def delete_task_files_on_delete_task(instance, **kwargs):
+    shutil.rmtree(instance.get_task_dirname(), ignore_errors=True)
+
+
+@receiver(post_delete, sender=Data, dispatch_uid="delete_data_files_on_delete_data")
+def delete_data_files_on_delete_data(instance, **kwargs):
+    shutil.rmtree(instance.get_data_dirname(), ignore_errors=True)